from functino to api conversion
