+++ .gitignore
@@ -1,0 +1,3 @@
+venv
+package-lock.json
+node_modules
